api/pyproject.toml
[tool.poetry]
name = "shopping_list"
version = "0.0.0"
description = ""
authors = ["Marek Zelazny"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.73.0"
alembic = "^1.7.5"
psycopg2-binary = "^2.9.2"
SQLAlchemy = "^1.4.27"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
email-validator = "^1.1.3"
uvicorn = "^0.17.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.9.3"
flake8 = "^4.0.1"
flake8-eradicate = "^1.2.0"
flake8-comprehensions = "^3.8.0"
flake8-bugbear = "^22.1.11"
pytest = "6.2.5"
freezegun = "^1.1.0"
requests = "^2.26.0"
ipython = "^8.0.1"

[tool.black]
target-version = ["py310"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
